requirements.txt
fastapi==0.85.1
motor==3.0.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
sympy==1.11.1
uvicorn==0.19.0
google-auth==2.14.1
jwt==1.3.1
itsdangerous==2.1.2
authlib==1.1.0
httpx==0.23.0
pyjwt==2.6.0
boto3==1.26.14
python-multipart==0.0.5